+++ CHANGELOG.md
@@ -1,13 +1,30 @@
 # Changelog
 
-# 2.3.11 (February 26, 2024)
+# 2.4.0 (March xx, 2024)
 
 ## Notes
 - Golang: **v1.19**
 
-### Enhacements
+### ZPA Additions
+
+- Added `POST /emergencyAccess/user` to create an emergency acess user for the specified customer. [PR #226](https://github.com/zscaler/zscaler-sdk-go/pull/226) - Added support to ZPA 
+- Added `GET /emergencyAccess/user` to get all emergency acess users for the specified customer. [PR #226](https://github.com/zscaler/zscaler-sdk-go/pull/226) - Added support to ZPA 
+- Added `GET /emergencyAccess/user/{userId}` to get the emergency access user for the specified customer. [PR #226](https://github.com/zscaler/zscaler-sdk-go/pull/226) - Added support to ZPA 
+- Added `PUT /emergencyAccess/user/{userId}`to update the emergency access user for the specified customer. [PR #226](https://github.com/zscaler/zscaler-sdk-go/pull/226) - Added support to ZPA 
+- Added `PUT /emergencyAccess/user/{userId}/activate` to activate the emergency access user for the specified customer. [PR #226](https://github.com/zscaler/zscaler-sdk-go/pull/226) - Added support to ZPA 
+- Added `PUT /emergencyAccess/user/{userId}/deactivate` to deactivate the emergency access user for the specified customer. [PR #226](https://github.com/zscaler/zscaler-sdk-go/pull/226) - Added support to ZPA 
+
+- Added `POST and PUT /mgmtconfig/v2/admin/customers/{customerId}/policySet/{policySetId}/rule` endpoints for access policy rule creation. This endpoint allows for larger payload submission. [PR #228](https://github.com/zscaler/zscaler-sdk-go/pull/228) - Added support to ZPA 
+
+# 2.3.11 (February 28, 2024)
+
+## Notes
+- Golang: **v1.19**
+
+### Fixes
 
 - [PR #230](https://github.com/zscaler/zscaler-sdk-go/pull/230) - Implemented centralized sweep for ZIA and ZPA packages.
+- [PR #231](https://github.com/zscaler/zscaler-sdk-go/pull/231) - Fixed ZPA Application Segment PRA changed attribute change from `sraPortal` to `praApps`.
 
 # 2.3.10 (February 22, 2024)